Move: 검색 페이지 컴포넌트 파일 이동
