+++ src/app/(provider)/(root)/category/search/_components/RecentSearch.tsx
