+++ src/app/(provider)/(root)/category/search/_components/SearchInput.tsx
