+++ src/app/(provider)/(root)/category/search/_components/index.ts
@@ -1,0 +1,2 @@
+export { default as RecentSearch } from './RecentSearch';
+export { default as SearchInput } from './SearchInput';
